Type improvements.

+++ src/extra.cpp
@@ -104,7 +104,7 @@ void ALS_multicore(smat_t &R, mat_t &W, mat_t &H, parameter &param){
 
         //optimize W over H
 #pragma omp parallel for schedule(kind)
-        for (int Rw = 0; Rw < R.rows; ++Rw){
+        for (long Rw = 0; Rw < R.rows; ++Rw){
             float *Wr = &W[Rw][0];
             int omegaSize = R.row_ptr[Rw + 1] - R.row_ptr[Rw];
             float ** subMatrix;
@@ -119,7 +119,7 @@ void ALS_multicore(smat_t &R, mat_t &W, mat_t &H, parameter &param){
                 //a trick to avoid malloc
                 float ** H_Omega = (float **)malloc(omegaSize * sizeof(float *));
                 unsigned i = 0;
-                for (int idx = R.row_ptr[Rw]; idx < R.row_ptr[Rw + 1]; ++idx){
+                for (unsigned idx = R.row_ptr[Rw]; idx < R.row_ptr[Rw + 1]; ++idx){
                     H_Omega[i] = &H[R.col_idx[idx]][0];
                     ++i;
                 }
@@ -138,7 +138,7 @@ void ALS_multicore(smat_t &R, mat_t &W, mat_t &H, parameter &param){
                 //sparse multiplication
                 for (int c=0; c < k; ++c){
                     subVector[c] = 0;
-                    for (int idx = R.row_ptr[Rw]; idx < R.row_ptr[Rw + 1]; ++idx){
+                    for (unsigned idx = R.row_ptr[Rw]; idx < R.row_ptr[Rw + 1]; ++idx){
                         unsigned idx2 = R.colMajored_sparse_idx[idx];
                         subVector[c] += R.val[idx2] * H[R.col_idx[idx]][c];
                     }
@@ -171,7 +171,7 @@ void ALS_multicore(smat_t &R, mat_t &W, mat_t &H, parameter &param){
 
         //optimize H over W
 #pragma omp parallel for schedule(kind)
-        for (int Rh = 0; Rh < R.cols; ++Rh){
+        for (long Rh = 0; Rh < R.cols; ++Rh){
             float *Hr = &H[Rh][0];
             unsigned omegaSize = R.col_ptr[Rh + 1] - R.col_ptr[Rh];
             float ** subMatrix;
